Initial public commit of corrupt profile scanner.
Get-CorruptProfiles runs a Test-Path on C:\Users\ to check for users
with a .domain (in this case whatever your domain is) suffix. A .domain,
.domain.000, .domain.000.001 or so on typically indicates a corrupt
profile on machines whose user profiles are changed or deleted
often--perhaps some type of kiosk, public access, or student lab
machines. If *.domain* profiles are found Get-CorruptProfiles looks
through Users selects Name, and PSComputerName properties allowing for
easy export to csv, json, or other filetypes.

Do note this does not remove profiles, it just provides a list of
potentially corrupt profiles and computers on which they reside.